Create LAB 4.4
By using logical operators I have created a program that a user inputs there year as well as following up with the GPA. If your GPA is 2.0 or greater then an message will pop up stating "you are doing well, the road to graduation is near!" with a if else if statement. This program is specified for seniors!