Create LAB 4.6
This program takes 4 quarters of the monthly water bill and averages out by 12 months in a year. After It has calculated the average monthly bill, It goes through the if else if statement to see weather or not you have used an excessive amount of water. But if the average is bigger than 25 and less then 75. An message will pop up basically stating that you have not used the typical amount of water & thank you for using a very small amount of water. " Thanks for saving the water".